Adding CSP Nonce Support
In the event a site is using a Content Security Policy to lock down what styles and scripts can be used, the Polisher class was not including a nonce for any style elements it creates.  This results in those style elements being ignored by the site for violating its CSP.

The Polisher will now look for the presence of a "csp-nonce" meta element and extract the nonce to use from that.  Do to an upcoming change to how that should be stored, it will first attempt to reference the `nonce` attribute on the meta element (the new location for the nonce) before falling back to the `content` attribute (the current location for the nonce).  If found, this nonce will be added to any style element created by the Polisher.