Fixing the use of `this.cspNonce` instead to reduce the number of lookups
I hadn't caught the first time the use of `insert()` during the setup, so I had tried to remove the use of `this.cspNonce` for a local variable, although I'd not removed `this.` from all locations.  However, since we can just do the lookup first on the setup, we can properly use `this.cspNonce` and only perform the lookup once.

I moved the lookup to first in the setup, then restored the use of `this.cspNonce` and removed the extra lookup during `insert()`.