Replacing the nonce lookup with the ability to pass in the nonce value to use
The Polisher will now accept a second parameter, called `cspNonce` which is immediately set to `this.cspNonce` in the constructor.  I removed the lookup and reference to the lookup function from the object.

The Previewer will now pass the value of `this.settings.cspNonce` into the Polisher when it instantiates it.  This allows a developer to pass the nonce value into the options object for the Previewer instantiation.  If using the polyfill, this is also settable in `window.PagedConfig` under the `settings` config option.

I added a section to the README to document the new options and how to set the value for both the NPM module and polyfill.